# license-application-service/src/main/resources/api.yaml
openapi: 3.1.0
info:
  title: Mallorca Holiday Rental Licensing API
  version: 1.0.0
  description: |
    API for managing applications, licenses, and payments for the Mallorca Holiday Rental Licensing system.
    Supports GDPR-compliant user management through Keycloak integration and ensures transparent, fair allocation of ETV licenses.
servers:
  - url: https://localhost:8080/
...
security:
  - KeycloakOAuth2:
      - openid
      - profile
      - email
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      description: Retrieve a paginated list of users, optionally filtered by username or email, supporting search and pagination parameters.
      operationId: listUsers
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: username
          in: query
          schema:
            type: string
          description: Filter users by username.
          example: johndoe
        - name: email
          in: query
          schema:
            type: string
            format: email
          description: Filter users by email.
          example: johndoe@example.com
        - name: first
          in: query
          schema:
            type: integer
            minimum: 0
          description: Index of the first result to return (pagination offset).
          example: 0
        - name: max
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Maximum number of results to return.
          example: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResource"
        "400":
          description: Invalid query parameter
        "401":
          description: Unauthorized

    post:
      summary: Create a new user
      description: Create a new user account with the provided details.
      operationId: createUser
      tags: [Users]
      security:
        - KeycloakOAuth2: [openid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                schema:
                  $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User created successfully
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "409":
          description: User already exists
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Retrieve detailed information for a specific user by their unique identifier.
      operationId: getUser
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResource"

    patch:
      summary: Update user attributes or password
      description: Update the attributes or password of a specific user.
      operationId: updateUser
      tags: [Users]
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                enabled:
                  type: boolean
                credentials:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [password]
                      value:
                        type: string
                      temporary:
                        type: boolean
      responses:
        "204":
          description: User updated successfully
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: User not found

    delete:
      tags: [Users]
      summary: Delete a user by ID
      description: Remove a user from the system by their ID.
      operationId: deleteUser
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "204":
          description: User deleted
        "404":
          description: User not found
  /users/{userId}/notifications:
    get:
      tags: [Users]
      summary: Return all notifications for a user.
      operationId: getNotifications
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of notifications for a user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserNotificationResource"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: User not found
  /users/{userId}/change-password:
    post:
      tags: [Users]
      summary: Change password for a user.
      description: Change the password for a specific user.
      operationId: changePasswordForUser
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  format: password
                  description: User's current password.
                  example: oldsecurepassword123
                new_password:
                  type: string
                  format: password
                  description: User's new desired password.
                  example: newsecurepassword456
      responses:
        "201":
          description: Password changed successfully
        "400":
          description: Bad request - invalid input data
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - incorrect old password
        "404":
          description: User not found
  /users/notifications/{id}:
    patch:
      tags: [Users]
      summary: Update status of a notification.
      description: Update the read status of a specific notification for a user.
      operationId: updateNotification
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated notification status.
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Notification not found
  /users/{userId}/notifications/preferences:
    get:
      tags: [Users]
      summary: Get notification preferences for a user
      description: Retrieve the notification preferences for a specific user.
      operationId: getNotificationPreferences
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResource"
    patch:
      tags: [Users]
      summary: Update notification preferences for a user
      description: Update the notification preferences for a specific user.
      operationId: updateNotificationPreferences
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferencesUpdate"
      responses:
        "200":
          description: Updated notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResource"
...
components:
  securitySchemes:
    KeycloakOAuth2:
      type: oauth2
      description: OAuth2 with Keycloak JWT Bearer tokens.
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/auth/realms/license-realm/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/auth/realms/license-realm/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: Profile information access
            email: Email access
  schemas:
...
    UserResource:
      type: object
      description: Represents a user of the holiday rental licensing portal, typically sourced from a Keycloak identity provider.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the user.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        username:
          type: string
          description: Unique username for login.
          example: johndoe
        email:
          type: string
          format: email
          description: User's email address.
          example: johndoe@example.com
        firstName:
          type: string
          description: User's first name.
          example: John
        lastName:
          type: string
          description: User's last name.
          example: Doe
        enabled:
          type: boolean
          description: Indicates if account is active.
          example: true
        emailVerified:
          type: boolean
          description: Whether the email has been verified.
          example: true
        createdTimestamp:
          type: integer
          format: int64
          description: Time the user was created (milliseconds since epoch).
          example: 1699458000000
      required:
        - username
        - enabled
    UserCreate:
      type: object
      description: Request payload for creating a new user in Keycloak.
      properties:
        username:
          type: string
          description: Unique username for login.
          example: johndoe
        email:
          type: string
          format: email
          description: User's email address.
          example: johndoe@example.com
        firstName:
          type: string
          description: User's first name.
          example: John
        lastName:
          type: string
          description: User's last name.
          example: Doe
        enabled:
          type: boolean
          description: Indicates if account is active.
          default: true
          example: true
        credentials:
          type: array
          description: List of user credentials.
          items:
            type: object
            properties:
              type:
                type: string
                enum: [password]
                description: Credential type.
                example: password
              temporary:
                type: boolean
                description: Whether this password is temporary.
                default: false
                example: false
              value:
                type: string
                description: Password value (hashed or direct depending on Keycloak configuration).
      required:
        - username
        - enabled
    UserNotificationResource:
      type: object
      description: A notification for a user
      properties:
        id:
          type: string
          format: uuid
          description: Unique message identifier.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        application_id:
          type: integer
          description: Unique application identifier.
          example: 201
        user_id:
          type: string
          format: uuid
          description: ID of the applicant user.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        date:
          type: string
          format: date-time
          description: Event time of the notification.
          example: 2025-05-20T08:00:00Z
        message:
          type: string
          description: The message of the notification.
          example: "Your application has been approved."
        is_read:
          type: boolean
          description: Status if the message has been read by the user.
      required:
        - id
        - user_id
        - date
        - message
        - read
...
    NotificationPreferencesResource:
      type: object
      description: User notification preferences for the holiday rental licensing portal.
      properties:
        id:
          type: integer
          description: Unique identifier of the notification preferences record.
          example: 301
        user_id:
          type: string
          format: uuid
          description: Unique identifier of the user.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        notification_way:
          $ref: "#/components/schemas/NotificationWayApiEnum"
        application_updates_notification:
          type: boolean
          description: Whether to receive notifications about application updates.
          example: true
        license_renewal_notification:
          type: boolean
          description: Whether to receive notifications about license renewals.
          example: false
      required:
        - id
        - user_id
        - notification_way
        - application_updates_notification
        - license_renewal_notification
    NotificationPreferencesUpdate:
      type: object
      description: Payload for updating user notification preferences.
      properties:
        notification_way:
          $ref: "#/components/schemas/NotificationWayApiEnum"
        application_updates_notification:
          type: boolean
          description: Whether to receive notifications about application updates.
          example: true
        license_renewal_notification:
          type: boolean
          description: Whether to receive notifications about license renewals.
          example: false
    NotificationWayApiEnum:
      type: string
      description: Methods of notification delivery.
      enum:
        - EMAIL # Notifications via email
        - SMS # Notifications via SMS
        - NONE # No notifications